keep-alive: limit wal_end to the oldest transaction still being processed
Currently we always return the server's wal_end+1
This approach has a few limits, can cause some issues.

If we start processing wal_end A,
then receive a keep-alive request with wal_end=A+x and reply A+x+1,
this before A if finished processing,
then if our process crashes the server will restart at A+x+1 and
will never receive it again.

To avoid this, this PR returns wal_end+1 only if there are no processing transactions
if there are it returns the wal_end of the oldest transaction still in progress.

This way we will be guaranteed to start back from there(A and not A+x+1) if we crash.